move buttons into component
