+++ src/components/About/About.svelte
@@ -6,7 +6,7 @@
   import AboutGallery from './AboutGallery.svelte';
 </script>
 
-<div class="f-center flex-col w-full gap-[80px] xl:gap-[120px] container">
+<div class="f-cl w-full container">
   <!-- About Gallery -->
   <AboutGallery />
   <!-- Factions Gallery -->
